C++Basics
REMEMBER to include that crucial semicolon at the end of any method and declare const variables! ALSO important thing to remember is that in a header file, the last bracket must be enclosed with a semicolon while the cpp file (where basically everything else happens) doesn't have as much semicolons to worry about in awkward places.